feat: Zoom 모달 기능 추가 및 Calendar 클래스에서 Zoom 처리 로직 통합
